iconprops to types.ts
