+++ libs/island-ui/core/src/index.ts
@@ -74,6 +74,7 @@ export * from './lib/Link/ArrowLink/ArrowLink'
 export { IconDeprecated } from './lib/Icon/Icon'
 export * from './lib/Icon/IconTypes'
 export * from './lib/IconRC/Icon'
+export * from './lib/IconRC/types'
 export * from './lib/Typography/Typography'
 export * from './lib/Text/Text'
 export * from './lib/Hyphen/Hyphen'
