+++ libs/island-ui/core/src/lib/AlertBanner/AlertBanner.tsx
@@ -3,7 +3,8 @@ import { Colors } from '@island.is/island-ui/theme'
 
 import { Box } from '../Box/Box'
 import * as styles from './AlertBanner.css'
-import { Icon, IconMapIcon } from '../IconRC/Icon'
+import { Icon } from '../IconRC/Icon'
+import { IconMapIcon } from '../IconRC/types'
 import { Text } from '../Text/Text'
 import { LinkContext } from '../context/LinkContext/LinkContext'
 import { Link } from '../Link/Link'
